Revert adding new class
